refactor(token-vault): refactor-token-vault-to-single-package
refactor token vault to a single package to simplify the build process
