+++ packages/token-vault/src/index.ts
@@ -1,4 +1,4 @@
-import { interceptor as interceptorModule } from '@shared/workers';
+import { interceptor as interceptorModule } from './lib/worker';
 import { client as clientModule } from './lib/client';
 import { proxy as proxyModule } from './lib/proxy';
