+++ packages/token-vault/src/lib/types/worker.types.ts
@@ -1,4 +1,4 @@
-import type { BaseConfig, ForgeRockConfig } from '@shared/types';
+import type { BaseConfig, ForgeRockConfig } from './config.types';
 
 export interface InterceptorConfig {
   events?: BaseConfig['events'];
