+++ packages/token-vault/src/lib/worker/index.ts
@@ -1,0 +1,2 @@
+export * from './interceptor';
+export * from './worker.utilities';
