--- packages/token-vault/tsup.config.ts
@@ -1,11 +1,0 @@
-import { defineConfig } from 'tsup';
-
-export default defineConfig({
-  entry: ['packages/token-vault/src/index.ts'],
-  splitting: false,
-  sourcemap: true,
-  dts: false,
-  clean: false,
-  format: ['cjs', 'esm'],
-  outDir: 'dist/packages/token-vault',
-});
