--- shared/network/README.md
@@ -1,11 +1,0 @@
-# shared-network
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build shared-network` to build the library.
-
-## Running unit tests
-
-Run `nx test shared-network` to execute the unit tests via [Jest](https://jestjs.io).
