--- shared/network/project.json
@@ -1,49 +1,0 @@
-{
-  "name": "shared-network",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "shared/network/src",
-  "projectType": "library",
-  "targets": {
-    "build": {
-      "executor": "@nx/js:tsc",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "outputPath": "dist/shared/network",
-        "main": "shared/network/src/index.ts",
-        "tsConfig": "shared/network/tsconfig.lib.json",
-        "assets": ["shared/network/*.md"],
-        "transformers": [
-          {
-            "name": "typescript-transform-paths/nx-transformer"
-          },
-          {
-            "name": "typescript-transform-paths/nx-transformer",
-            "options": { "afterDeclarations": true }
-          }
-        ]
-      }
-    },
-    "lint": {
-      "executor": "@nx/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["shared/network/**/*.ts"]
-      }
-    },
-    "test": {
-      "executor": "@nx/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-      "options": {
-        "jestConfig": "shared/network/jest.config.ts",
-        "passWithNoTests": true
-      },
-      "configurations": {
-        "ci": {
-          "ci": true,
-          "codeCoverage": true
-        }
-      }
-    }
-  },
-  "tags": ["scope:lib"]
-}
