--- shared/network/src/index.ts
@@ -1,2 +1,0 @@
-export * from './lib/network.utilities';
-export * from './lib/network.types';
