--- shared/types/README.md
@@ -1,7 +1,0 @@
-# shared-types
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build shared-types` to build the library.
