--- shared/types/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "@shared/types",
-  "private": true,
-  "version": "0.0.0",
-  "type": "commonjs",
-  "main": "./src/index.js",
-  "exports": {
-    ".": "./index.js"
-  }
-}
