--- shared/types/project.json
@@ -1,35 +1,0 @@
-{
-  "name": "shared-types",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "shared/types/src",
-  "projectType": "library",
-  "targets": {
-    "build": {
-      "executor": "@nx/js:tsc",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "outputPath": "dist/shared/types",
-        "main": "shared/types/src/index.ts",
-        "tsConfig": "shared/types/tsconfig.lib.json",
-        "assets": ["shared/types/*.md"],
-        "transformers": [
-          {
-            "name": "typescript-transform-paths/nx-transformer"
-          },
-          {
-            "name": "typescript-transform-paths/nx-transformer",
-            "options": { "afterDeclarations": true }
-          }
-        ]
-      }
-    },
-    "lint": {
-      "executor": "@nx/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["shared/types/**/*.ts"]
-      }
-    }
-  },
-  "tags": ["scope:types"]
-}
