--- shared/types/src/index.ts
@@ -1,2 +1,0 @@
-export * from './lib/config.types';
-export * from './lib/tokens.types';
