--- shared/workers/README.md
@@ -1,11 +1,0 @@
-# shared-workers
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build shared-workers` to build the library.
-
-## Running unit tests
-
-Run `nx test shared-workers` to execute the unit tests via [Jest](https://jestjs.io).
