--- shared/workers/jest.config.ts
@@ -1,10 +1,0 @@
-/* eslint-disable */
-export default {
-  displayName: 'shared-workers',
-  preset: '../../jest.preset.js',
-  transform: {
-    '^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
-  },
-  moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '<rootDir>coverage/shared/workers',
-};
