--- shared/workers/src/index.ts
@@ -1,3 +1,0 @@
-export * from './lib/interceptor';
-export * from './lib/worker.types';
-export * from './lib/worker.utilities';
