+++ tsconfig.base.json
@@ -8,10 +8,7 @@
       "@forgerock/javascript-sdk": ["packages/javascript-sdk/src/index.ts"],
       "@forgerock/ping-protect": ["packages/ping-protect/src/index.ts"],
       "@forgerock/token-vault": ["packages/token-vault/src/index.ts"],
-      "@plugins/wait-for-api": ["plugins/wait-for-api/src/index.ts"],
-      "@shared/network": ["shared/network/src/index.ts"],
-      "@shared/types": ["shared/types/src/index.ts"],
-      "@shared/workers": ["shared/workers/src/index.ts"]
+      "@plugins/wait-for-api": ["plugins/wait-for-api/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
